IBP: Fixed provision fail on a node with one disk
To address some bugs about large amount of disks,
huge disks, etc. we introduced workarounds in
nailgun data driver in fuel agent.

The issue is that when only one disk on a node,
nailgun places /boot not on raid but on plain parition
and flag boot_done is set only inside 'if raid' clause,
but not inside 'if partition' clause.

Change-Id: I8a598ccde661fae375bb102f36550fb53f1abbe8
Closes-Bug: #1465971