Nodes metadata filtration added
Filtering of private nodes metadata added to statsender.
This patchset is only for syncronization of code in stable/6.1 and master.
Current whitelists implementation is not fit to fix this issue due to
meatdata data structure. Filtration by whitelists is going to be implemened
in: https://bugs.launchpad.net/fuel/+bug/1466121

Change-Id: I2c5f39e0ac0f77c281db444933ac54723c2645c7
Closes-Bug: #1466140