Use schemaless approach for node roles
In Fuel 7.0 we want to have ability to extend node roles on fly by using
plugins. In order to achieve this, we want to have declarative
schemaless approach, so we won't require database object manipulations
(CRUD) each time we somehow change node role.

This commit drops such tables like:

* roles
* pending_roles
* node_roles

and use JSON arrays in nodes tables instead:

* roles
* pending_roles
* primary_roles

Implements blueprint role-as-a-plugin

Change-Id: I2d3fccb7187fa4e82595c9ca6b27ef8ce90ee4b2
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>