Each extension should have name and version
Combination of name and version will be used
as uniq identificator of extension. This identificator
is required to link node to specific extension,
assign uniq name for version table which is required
for alembic migrations and provide extension version
specific url in order to prevent collisions
between different extensions.

Implements blueprint: volume-manager-refactoring

Change-Id: I92566b32c44440df4bc45e0cddf2d66ac9156a5f