Get rid of dependencies on fixed networks' names
We cannot rely on names of networks anymore as netorking with templates
will allow to create/delete networks and set (almost) arbitrary names
to them. So, we need to tie services for network roles.
While network roles descriptions are in development we can use hardcoded
version where network roles are in plain list and network roles to networks
mapping is fixed.

Change-Id: Ibf523885faa57bb2435f9bf6386258b69063b318
Implements: blueprint templates-for-networking