IBP: fuel-agent should preserve free space in produced images
As for now, fuel-agent produces images which were shrunk to the minimum
filesystem size. Therefore, it could result in a file system which is not
optimized for speed.

This change will allow to control the amount of preserved free space in
produced images.

Additionally, it enlarges /boot partition size to 300M in order to fit newly
produced images into /boot partition size.

Change-Id: I8a2fd71f33b15de49b2854b03672e1c2eec4979d
Closes-Bug: #1469155