Add processing of keep_data flag
In case when some partition has keep_data = True
need to elevate flag to file system level and
do not to create new partition table. Do not clean
file system when it has keep_data = True.
There should not be a problem if some raw data
will not be cleaned.

Change-Id: I9e67e4b7f9421d517e360e403fa89ea8ca9f43b8
Implements: blueprint partition-preservation