Add generators for uuid and cephx
In order to become more advanced with the ceph modules, we need to more
precisely create some of the attibutes like the cluster uuid or the user
cephx auth keys. We need to add the generators so that the modules can be
worked on outside of the release schedule and proposed again later.

Change-Id: I82f917cdaf5053fc45cf6071f598fb82da43846a
Closes-bug: #1475759