Add generators for uuid and cephx
In order to become more advanced with the ceph modules, we need to more
precisely create some of the attributes like the cluster uuid or the user
cephx auth keys.

While we don't use these generators now, adding them is necessary to that
parallel development work can occur on the library side. Updates to the
fixtures will be proposed when the library modules are ready to integrate.

Change-Id: I82f917cdaf5053fc45cf6071f598fb82da43846a
Closes-bug: #1475759