Add timeout value to 'urlaccesscheck check'
This change adds a configurable timeout to 'urlaccesscheck check' to
ensure the script does not hang if there are connectivity issues with
the URL that is being checked.  The default timeout is 60 seconds but
is configurable via a command line option.

This change also adds netifaces to the network_checker requirements.txt
because it was not added as part of
I910c15c2b39a917eb8428bb69271b5dde364b639 so tests fail on a clean
environment.

Change-Id: I363712a6501cc21b89601d5373b5633d5518ca81
Closes-Bug: 1478138