Necessary extending of objects to perform upgrade
The upgrade procedure should create a new cluster with the same
parameters as the original cluster. The new method was intoduced to
parameters the cluster that was created with. The return value of the
new method is compatible with the data that are expected by the create
method.

A set of hooks of the extension mechanism was extended by the
fire_callback_on_cluster_delete hook. This hooks is fired when cluster
is deleted.

Two opposite method were added: get_assigned_vips and
assign_given_vips_for_net_groups. The first method returns all assigned
VIPs of the cluster and the second complementary method receives the
same format that the first returns to assign given VIPs.

The patch has some impact on the get_admin_network_group method. This
impact consist of adding a new filter called node_group_id that allows
to filter network groups by ID of the node group when there are no any
nodes in the cluster.

Implements blueprint: nailgun-api-env-upgrade-extensions

Change-Id: I156d3c71a7e832a2eb99a1058a6ac628fab71483