Don't rely on networks mapping stored in DB for template
Method get_node_network_by_netname() uses networks-to-interfaces mapping
stored in DB. So, this method cannot be used in template serializer as
this data in DB is not relevant in case of using template.

Another approach should be used to get IP information for node's
networks. So, get_node_networks_ips() method is added to get IP information
on the base of nodes-to-networks mapping which is relevant when template
is loaded.

Change-Id: I16757bee15a6151a815f13d427ee9447e6e1b0a9
Closes-Bug: #1477953