Allow changing/deletion of template for any cluster state
Initially, it was denied for cluster when it is not in new or stopped state
as for other network configuration. But that leads to usability issues
described in ticket. It is to be clarified in docs that user should ensure
that the changes in configuration are syncronized on all nodes.

DocImpact
Change-Id: I044f7c2cecc5b4c9561f0cb0c1da887e58c96651
Closes-Bug: #1477647