Remove regex validation for 'target_node' attribute
'target_node' is an object which gets transmitted from frontend.  It is
a python dictionary, not a string, but in attribute definition it has a
'regex' attribute, which forces validation for value contained in
'target_node'.  Validation tries to validate string, but receives
dictionary on input and backend produces a traceback (TypeError:
expected string or buffer).  Fix problem by removing 'regex' validation
for 'target_node'.

Change-Id: I465e8a70ca5ea287f65e36cc3492691aacad395b
Closes-bug: #1479293