Preserve original order even if all dependencies are skipped
It was possible to lose all edges when if subgraph was created
with skipped tasks.
For example, if we have path a -> b -> c.
And b will be marked as skipped - there is no way to guarantee
that a will be executed earlier than c.

After this patch traversal will be done using all tasks and
after that filtered tasks will be skipped from sending them
to executor or api.

Change-Id: Iab95774b913c6c4dcd44b5213823e895f1c1f272
Closes-Bug: 1479829