Allow '_' in node interface names
Since it's now possible to assign arbitrary network interface names,
the underlying code that creates udev rules must support this.
Fuel agent expects mapping of MAC addresses to interface names to
be in the following format: <mac>_<ifname> and splits the entry
without limiting the number of splits. If <ifname> contains '_'
this code will produce 'ValueError: too many values to unpack' error.
This patch fixes this behaviour by limiting the number of splits to 1.

Change-Id: Ie7043be8b2bfc5025d9563eb97157f73a8eb1296