Decrease amount of 'get_node_networks` calls
We had a lot of 'get_node_networks' calls for the same node. This
function call is too expensive and takes a lot of time. It cause some
serious problems on scale, since its call number was increasing with
a progression. Since now, it's predictable and takes only 3 * N times.

Here's the quick comparison:

  NODES | WAS | NOW
  ----------------
    6   | 216 | 18        // was: 6.88s, now: 0.62s
    5   | 155 | 15
    4   | 104 | 12
    3   |  63 |  9

Closes-Bug: #1479090

Change-Id: Ia8ea44149fbb3c90aa015ad5ac0f03f4cd9d33ec
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>