Properly update IP ranges for net groups by CIDR
NetworkManager.update_range_mask_from_cidr ignores a gateway if it is
used in a network group. The patch adds the use_gateway parameter to the
function to properly select the first address of a range.

NetworkManager.update_networks does not check a notation of IP ranges
properly and due to the fact that the order of an iteration through
items depends on implementation of the dict type in Python we have that
the _set_ip_ranges method can occure after the
update_range_mask_from_cidr method and overrides proper ranges. The
patch removes logic for choosing an appropriate method from the loop.

Closes-Bug: #1480228
Co-Authored-By: Artur Svechnikov <asvechnikov@mirantis.com>
Change-Id: I46aff72d3f1c395f2bb2f82cdcb60b132eeef3c9