Use network roles data for assigning of VIPs
Additional information from VIPs description should be included into
VIPs data serialized for orchestrator in 7.0. It was correct in case of
using Neutron but it was not changed for Nova-Network case sinse 6.1.
NovaNetworkManager70 is introduced now and it provides the same VIPs
allocation as for NeutronManaget70 but without templates support.

Change-Id: Iddb1bb894e72084224a76d7de86c3ad57cf06b25
Closes-Bug: #1480202