Pin requirements to versions from Kilo
Updated version of upper packages bounds to be more in line
with those from Kilo's global requirements.
This solved some of the problems with constantly breaking
releases of new OSt & oslo packages.
Unfortunately not all of them are in line still so we need pinning
for example inc ase of stevedore.

Change-Id: Iefa3628af5ae86414082b7100a671544b354e355
Partial-Bug: #1481212