Add IP assignment optimized for network template
IPs were allocated for every network on every node
regardless of network to nodes mapping which is set in template.
It led to excessive IP usage.
Now it is fixed so that IPs are allocated only for nodes where they are required.
And IP allocation is optimized for the case with network template and for other
cases partially.
A number of small fixes were made on the base of comments in earlier RCs.

IP allocation comparison in DB queries count (only for replaced part):
old version: 1 query per IP to allocate + 1 query per occupied IP in that range
new version: 1 queries per network (for all IPs) if (occupied IPs)+(IPs to allocate) <= MIN_IPS_PER_DB_QUERY
             2 queries per network otherwise (typically).
New version is much better for large numbers of nodes.
nodes |      queries count        |
count |   old   |   new           |
      |min| max |min|  max        |
    1 |  1|  1+N|  1|1+ceil(N/5)**|
   10 | 10| 10+N|  1|1+ceil(N/10) |
  100 |100|100+N|  1|1+ceil(N/100)|
where N is a number of occupied IPs.
** 5 here is value of MIN_IPS_PER_DB_QUERY.

Change-Id: I4479bcae9e80f7b96750f9fdb2868d6df164baf6
Closes-Bug: #1480345