Remove new supervisor configs during rollback
We have to clean-up new artifacts in case of rollback and we didn't do
this for new supervisor configs. That's why upgrade fails if there were
some rollback before - we're unable to manage docker containers because
new one are launched automatically when we switched to the new
supervisor folder which is intended to be empty.

Closes-Bug: #1484091

Change-Id: Ic491d04fc3037c748b403fe0964f3fbc8db5769d
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>