Add check to NodeDeletionTaskManager for controllers to be deployed
In case we remove controller from ha env we must redeploy others which
is done in the manager. Check added to list of retrieved for deployment
controllers in order to make sure they are in appropriate for the
operation status. In opposite case we might faced bugs in particular
situations, for example, when environment is reset and implicit
deployment is started on nodes which still are undergoing the process and
that leads to unexpected errorfull behaviour as such described in the bug

Change-Id: I099660db1f8cd9d179f377799231635865286d3a
Closes-Bug: #1432615