Add /api/master_node_uid endpoint
To fix the bug linked below we need an endpoint in Nailgun
which will help fuel-nailgun-agent detect if master node
was reinstalled. This CR adds endpoint exposing master node
UID which is unique for any Fuel installation. Now
fuel-nailgun-agent can monitor that endpoint and consider
master node to be reinstalled if the UID changes.

Related-Bug: #1431386
Change-Id: I519e52542b089b13ef287256c172ababb7a362ba