Update nodes.yaml and /etc/hosts on node removal
Currently Nailgun doesn't support declarative scale-down mechanism, and
therefore there's no way to trigger some task execution if some node has
been removed from the environment, and that's incredibly annoying for
plugin developers.

Plugin developers could come up with a workaround, and have a deal with
/etc/hiera/nodes.yaml. Unfortunately, this file isn't updated on node
removal event. This commit is going to fix it, and send two tasks on
execution if case of node removal:

* update /etc/hiera/nodes.yaml
* update /etc/hosts

Closes-Bug: #1475363
Closes-Bug: #1491027

Change-Id: Ie9ecc5cd1b7cc185a1fc694e728d8ecc571df23d
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>