Refactor network availability for node check
Code of NetAssignmentValidator.validate updated with additional check
which determines whether networks in input data can be assigned to node.
The decision is made upon availability of particular net for node group
in which the node is included. The validation is meant to substitute
another check that does the same yet uses already assigned to the node's
nics networks which has produced errorful behavior in some cases (see
the bug's description)

Change-Id: I028b50e32d7bf1b9d02674b861e2e5ff24b4f2d5
Closes-bug: #1425907