Added MAINTAINERS to top-level repo directory
In a consensus reached in a proposal for code review process
change [1], and proposed team structure policy document [2],
adding MAINTAINERS files in repositories. This commit
adds the first file at the top level of fuel-web repository,
similar files to subfolders of the repo are to be committed
separately.
Please note, that files are written in YAML format in order
to allow further automation around code review process.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-August/072406.html
[2] http://lists.openstack.org/pipermail/openstack-dev/2015-September/074984.html
[3] https://review.openstack.org/#/c/225376/1

Change-Id: Ic5b7372ba083eb432cf98a9048fc53323a2af3d4