Added MAINTAINERS to fuelmenu
In a consensus reached in a proposal for code review process
change [1], and proposed team structure policy document [2],
adding MAINTAINERS files in repositories.
This commit adds a maintainer to fuelmenu in fuel-web repo.

Please note, that files are written in YAML format in order
to allow further automation around code review process.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-August/072406.html
[2] http://lists.openstack.org/pipermail/openstack-dev/2015-September/074984.html
[3] https://review.openstack.org/#/c/225376/1

Change-Id: If82536d7192062cbd2871391b7b02e40187b8bb5