ErrorsFactory returns only defined errors
ErrorsFactory had __getattr__ method that was used to always
return any error (if it doesn't existed, it created a new one).
This was causing "hasattr" to return 'True' for every check.
Due to this any code, that wanted to check for some unique attributes
in ErrorsFactory was working incorrectly. This was the case for
a py.test library when scannig code when looking for tests.
Now this behavior has been removed. This uncovered new bugs
which was fixed.

Change-Id: I3c9be96dbcc535ea9c730d29b065479cc318fd9f
Closes-Bug: #1497664
(cherry picked from commit 9757292c3b7f2a8d955fbe2cc1673a2baac702e5)