Fix attempt to access not existed network group id key in input data
Code of NetworkGroupValidator.validate_network_group utilize 'id' key
(id of the network group being operated on) so its absence will
lead to KeyError exception when the data is not present in payload of the
request, e.g. when user updates certain attributes of the network group
using fuel-client. This patch fixes described problem.

Change-Id: Ida494738ce5d61a6eec20bfa815923c9da6a7d56
Closes-Bug: #1500308