Ditch Subs
Initially, Subs was introduced to substitute sensitive data from logs
which are getting included into diagnostic snapshot.
However, it was turned off by default since fuel 5.1

And we're in a situation where this piece of code was left untouched
for almost 2 year or so.

By getting rid of it, we is going to solve few issues:
1) devs will no longer need to spend time to maintain unused piece of
code.
2) shotgun will stop to modify the actual data inside of log
files. Therefore, this will allow us to significally reduce waste of
disk space and i/o on preparing copy of locally included logs.

Change-Id: Id78e369ea04e2847025c15150038de5f508a33af
Related-Bug: #1445311