Performance test timeout increased for PUT of DeploySelectedNodes
We have periodical failures of test_put_deploy_selected_nodes test
on the CI server. This failure is reproduced periodically in all
the time of job monitoring. Thus we can assume, that timeout value
is not optimal for our CI servers.
Thus acceptable timeout for put_deploy_selected_nodes increased
from 10 to 15.

Change-Id: I69f4e1b11394cbff110d473113574920a236f50b
Closes-Bug: #1498365