Add cloud management user
This change adds a new cloud_management section to the openstack.yaml.
In this change we are creating elements for an admin user, admin user
password and tenant that can be used for lifecycle related activities.
This will allow us to not require the normal admin account to perform
cloud operations.

The addition of a fuel-admin user would allow fuel to perform life cycle
related cloud actions even if the cloud adminstrator wants to change the
default admin password. Additionally, the cloud administrator could
disable the fuel-admin account to prevent fuel from being able to make
modifications but also allow them to enable it only when necessary to
make cloud changes.

DocImpact
Partial blueprint: compute-node-lifecycle

Change-Id: I5f4e5dfb4e6786d3e5b5d96e889fb2b771b68a97