Use appropriate types to store IP and MAC addrs
In PostgreSQL there are special types for storing IPv4, IPv6 address and
subnetworks and also MAC addresses. Storing data in these types is more
efficient than in strings because it requires less space. In addition,
it provides additional operations for columns of these types.

This patch changes types of some columns for models such as:

    IPAddr
        ip_addr: inet

    IPAddrRange
        first: inet
        last: inet

    Node
        mac: macaddr
        ip: inet

    NodeNICInterface
        mac: macaddr
        ip_addr: inet
        netmask: inet

    NodeNICInterface
        mac: macaddr

    NetworkGroup
        cidr: cidr
        gateway: inet

    NeutronConfig
        internal_cidr: cidr
        internal_gateway: inet

    NovaNetworkConfig
        fixed_networks_cidr: cidr

Co-Authored-By: Ilya Kharin <ikharin@mirantis.com>
Change-Id: Id8431eb7d2e3ca1a1cc29fc865d57359b29d0e35
Partial-Bug: #1484973