Replace release.is_deployable with release.state
is_deployable doesn't really give any additional information for a
release, it is removed, to make change API - compatible with API, it is
retained as a property dependent on release.state.

New state is added - manageonly, for environments that are not able to
be deployed, but can still be managed.

Change-Id: I518a0114730a2f227c9ef035a376f9a90d3d5bbd
Closes-bug: #1503303