Prepare `nailgun-net-check` package
Prepare `nailgun-net-check` package to upload it to `pypi`. It is
required for `fuelmenu`. So it will be added to the `requirements.txt`
when it will be available on `pypi`.

Move content of `setup.py` file to `setup.cfg` file.

Add `Makefile` for quick and handy `release` command.

Some changes (files) were taken from `openstack-dev/cookiecutter.git`.

Change-Id: Ia9119ac924f203cc814c986c9adb760baa461b90
Partial-Bug: #1502047