Check for existence of network group when uploading template
Network groups can be created at any time so there is no guarantee
that it will exist when a template is uploaded. Before updating the
network group's VLAN the existence of the network group is checked.

Networks are reassigned for all nodes when a new network group is
created. A new pre-deployment check was added to validate that all
network groups required by a network template exist.

Change-Id: I28d3df4db0fe69f5b4fc9af491f0aae1f695901f
Closes-bug: #1509543