Allow to fetch logs from the logrotated log files
Current implementation allows to fetch data only from such files as
`puppet-apply.log`. But for now there are also rotated log files (logs
which are created by logrotate) such as, just for example,
`puppet-apply.log.1.gz`. Which should be also parsed and returned in
response. The functionality, which is included in this patch, does this.

Closes-Bug: #1512280
Change-Id: I7dc2b330a95a3a97ed3b83925139135ed9cf4b1f