Rerun network setup on deployed nodes
This is required to apply changes that could be made to the networking
configuration of environment.
The major case is: add/remove node network group(s).

The configuration changes are not tracked as nailgun does not have
proper functionality for that.

New key is introduced in task description: reexecute_on (list).
Deployment tasks which are to be re-executed on deploy changes must have

reexecute_on:
  - deploy_changes

DocImpact

Change-Id: I668fcb426826cecf69a5c7fc1e92b8ae8dfeef9f
Blueprint: redeploy-network-on-changes