Set configurable flag for Admin network to True
Setting of this flag to True is required to get UI to show this network
on "Networks" tab. So, Admin networks which are tied to node network groups
became available for editing. Default Admin network remains hidden as
it is not tied to any node network group. It is expected behavior in this
ticket. Changing of default Admin network remains allowed via CLI only.

Change-Id: Ife33aca7848df80bcf47557835b988da1d43b04d
Closes-Bug: #1518342