Logging in finding disk function fixed
The funcion find_existing_disk is called in the VolumeManager constructor.
When the new node is discovered and nailgun-agent doesn't post yet nodes
info to the Nailgun we have no volumes info and have warning and error
messages in the logs on each VolumeManager object initialization.
On empty volumes data we shouldn't perform search in find_existing_disk.
Just log message and return None.

Change-Id: Ia7ffc5f69995dc4230f5984faaa08ae543bbe044
Closes-Bug: #1523471