Openstack configuration can not be updated in not ready nodes
Add check that configuration update can not be executed if
any of target nodes are not in 'ready' state.

Current solution is:

- If any deployment is running, then uploading new
configurations, deleting old ones and execution updates are prohibited
in order to keep cluster configuration in a consistent state

- By default Execute update can be done successfully, only if
cluster is in state 'operational' and *all* target nodes to update
are in the state 'ready'.

- 'force' flag allows to execute update configuration only on target
nodes which is currently in state 'ready'. Other target nodes will be
ignored. We assume that, these nodes nodes will get an updated
configuration during deployment.

Change-Id: I2ef06b7084b0cafa576172c9f2db40aa4ccd6db8
Closes-bug: #1523131