Openstack configuration can not be updated in not ready nodes
Add check that configuration update can not be executed if
any of target nodes are not in 'ready' state.

Current solution is:

- If some nodes present in deploying state, then uploading new
configurations, deleting old ones and execution updates are prohibited
in order to keep cluster configuration in a consistent state

- By default Execute update can be done successfully, only if
*all* nodes in the cluster are in the 'ready' state.

- Execution update with 'force' flag triggers update configuration
only on 'ready' nodes. We assume that, all other nodes will
get an updated configuration during deployment.

Change-Id: I2ef06b7084b0cafa576172c9f2db40aa4ccd6db8
Closes-bug: #1523131