Add 'identity' key to mcollective configuration data
'identity' parameter represents id of the node and is placed in
mcollective config by nailgun-agent. But sometimes such behavior
(especially if to take into consideration that restart of mcollective
follows it) may lead to failed deployment (See related bug). Now the parameter
is supplied by nailgun and is used by fuel-agent to create the config
with the data already present in it when node boots after provision is
done.

Change-Id: I753eb76ed9c3b80f249c0c4b86ef48ef49274990
Related-Bug: #1518306