Loading fixtures only once into DB implemented
On each start of the Nailgun container fixman.upload_fixtures is
called. If some object loaded from fixture is deleted we will try
to push this object from the fixture to the DB on conainer restart.
This caused by hardcoded ids in the fixture and checking that object
is loaded from fixture implementation. We can't remove hardcoded
ids from the fixtures for backward compatibility with old Fuel
installations.
As fix we are using Alembic migration workflow and Alembic migrations
branching feature. Now on fixman.upload_fixtures just tries to
apply Alembic migrations for branch 'fixtures'.

Change-Id: Ibcdbf55d16eb7d947058ba4ea68f692ad45b09a2
Closes-Bug: #1526917