Fixed executing plugins in pre deployment stage
Create repositories and sync tasks were added to the end of
pre deployment stage. As result environment was not ready for
executing task in pre deployment stage. Now in granular deployment
create repository and sync tasks adds to the begining of
pre_deployment stage. So plugins tasks can be executed in any point
of pre_deployment stage.

Change-Id: I0713904e6ae17763606630c649bf1263e6d1b37c
Closes-Bug: 1530351