Add 'once_deployed' flag to cluster that forbids stop deployment
New attribute added for cluster entitity - 'once_deployed' which value
is used as a base in decision wether stop deployment operation can be
performed for the cluster. When cluster is created the flag is 'false'
and is set when deployment has been sucessfuly completed that is the
cluster has 'operational' status.

The database migration changed so that already existing clusters
populates with 'once_deployed' = True if their status also is
'operational'.

Validator added for ClusterStopDeploymentHandler that now checks value
of the flag and throw corresponding error if it is True.

Change-Id: I3ada735c066db9f912f8f4bb6b5cef341a057977
Partial-Bug: #1529691