Data pipelines in extensions
Data pipelines in extensions for deployment and provisioning
serialized data.

Data pipeline is an attribute of Extension - a list of Classes
which provide one of (or both) methods: process_provisioning,
process_deployment. They can be used to change provisioning or
deployment data just before running the action.

Change-Id: Ie778c7ee6c9eb2abe2783b886ee6be481eccd2dc
Implements: blueprint data-pipeline