Reassign nodes without reinstallation
In some upgrade scenarios when shadow environments are used some of
nodes should not be reprovisioned during this procedure. It is useful in
combination when control plane nodes are reprovisioned and data plane
nodes are updated in place.

The update_cluster_assignment method of the Node class was changed to
accept roles and pending_roles for node during the reassignment. It
allows to specify proper roles by the upgrade extention. kernel_params
is preserved because there is no possibility to change this parameter
for already deployed nodes.

The NodeReassignHandler handler accepts two additional parameters:
  - reprovision = True (default) - allows to skip the reprovision step
  - roles = [] (default) - allows to specify new roles

Change-Id: Iedb20a904e58f5b9a86eb47de8d8d317dc3cc61b
Blueprint: upgrade-major-openstack-environment