Reassign nodes without reinstallation
In some upgrade scenarios when shadow environments are used some of
nodes should not be reprovisioned during this procedure. It is useful in
combination when control plane nodes are reprovisioned and data plane
nodes are updated in place.

The update_cluster_assignment method of the objects.Node class was
changed to accept roles and pending_roles for node during the
reassignment. It allows to specify proper roles by the upgrade
extention.

The NodeReassignHandler handler accepts two additional parameters in the
request body:
  - reprovision = True (default) - allows to skip the reprovision step
  - roles = [] (default) - allows to specify new roles or preserve the
                           current roles if empty

Two additional methods were added to NailgunClusterAdapter and
NailgunReleaseAdapter respectively.

Change-Id: Iedb20a904e58f5b9a86eb47de8d8d317dc3cc61b
Blueprint: upgrade-major-openstack-environment
Closes-Bug: #1558655